+++ NEWS.md
@@ -1,1 +1,8 @@
-* Hack to set pixel depth of the MikrotronMC408x camera.
+- Possibility to skip corrupted initial image(s).
+
+- Use `ScientificCameras` package and implement its API.
+
+- Add an access parameter in parameter/register definition so that it is easy to
+  figure out whether a parameter is readable or writable.
+
+- Hack to set pixel depth of the MikrotronMC408x camera.